refactor: extract shared recipe validation helpers
Move duplicated RECIPE_FIELDS, pickRecipeFields, and validateRecipeBody
from both route files into src/lib/recipe-validation.ts to keep POST
and PUT validation consistent.
